refactor(data-audit): decompose DataAuditService into Checks/Trace/Findings/Backfill services
Behavior-preserving, read-mostly (0 $tx; writes only to dataAuditLog). Facade
keeps the 24-method surface + both @Cron decorators + delegates. 12 invariant
checks → DataAuditChecksService; traceContract + 9 trace helpers → ContractTraceService;
findings/ack/SLA → AuditFindingsService; backfill (dead JournalAutoService dep) →
AuditBackfillService. tests green, tsc 0, lint 0.
